Fix form validation bug in CaptchaComponentComponent and add session storage feature to HomeComponentComponent
The code changes in CaptchaComponentComponent fix a bug where the form is not being validated properly. The `send` method now checks if the form is invalid and marks all controls as touched.

In HomeComponentComponent, a session storage feature has been added. The `currentStage` value is retrieved from session storage and used to set the initial value of `currentStage` in the `ngOnInit` method. Additionally, when the `onCaptchaSuccess` method is called, the `currentStage` is incremented and stored in session storage.